inference/requirements.txt
torch
torchvision
transformers
streamlit
easyocr
PyMuPDF
Pillow
numpy
scipy
protobuf